+++ src/mimic/orm/schema/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod build;
-pub mod helper;
 pub mod node;
 pub mod types;
 pub mod visit;