+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.1.8]
 - wrapped Db in an Rc so that it's easier to pass through to await code
 - added better fixture validation, it tells you the entity path now
+- introduced a schema Builder that allows you to register hooks pre-build/validate
 
 ## [0.1.7]
 - refactor of entire api crate with new error handling, and grouped submodules